Update the hashtable support hash search benchmark to use the new name convention for the methods and move the shared code to an external support c file
