+++ benches/bench-support.c
@@ -1,0 +1,24 @@
+#define _GNU_SOURCE
+#include <stdio.h>
+#include <stdlib.h>
+#include <sched.h>
+#include <pthread.h>
+
+#include "cpu.h"
+
+void set_thread_affinity(int thread_index) {
+#if !defined(__MINGW32__)
+    int res;
+    cpu_set_t cpuset;
+    pthread_t thread;
+
+    CPU_ZERO(&cpuset);
+    CPU_SET(thread_index % psnip_cpu_count(), &cpuset);
+
+    thread = pthread_self();
+    res = pthread_setaffinity_np(thread, sizeof(cpu_set_t), &cpuset);
+    if (res != 0) {
+        perror("pthread_setaffinity_np");
+    }
+#endif
+}
