+++ benches/bench-support.h
@@ -1,0 +1,14 @@
+#ifndef BENCH_SUPPORT_H
+#define BENCH_SUPPORT_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void set_thread_affinity(int thread_index);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif //BENCH_SUPPORT_H
